新增后台管理模块

+++ YCloud-core/src/main/java/cn/xiaosm/cloud/core/config/security/SecurityUtils.java
@@ -21,7 +21,7 @@ public class SecurityUtils extends DefaultSecurityUtils {
         }
     }
 
-    public static Integer getLoginUserId() {
+    public static Long getLoginUserId() {
         return SecurityUtils.getLoginUser().getId();
     }
